[updata] AB#1106 uploadFileToGCSを別ファイルに分ける
