+++ express-nodejs/src/index.ts
@@ -1,10 +1,9 @@
 import express from 'express'
-import { Storage } from '@google-cloud/storage'
 import Speech from '@google-cloud/speech'
-import { stringify, v4 as uuidv4 } from 'uuid'
 import sendgrid from '@sendgrid/mail'
 import { SecretManagerServiceClient } from '@google-cloud/secret-manager'
 import multer from 'multer'
+import { uploadFileToGCS } from '../src/GCS'
 
 namespace EnvironmentVariable {
     export let fromAddress: string = "";
@@ -28,25 +27,6 @@ getSecretManagerValue('send_email_address').then((result) => {
         console.error("emailアドレスの取得に失敗しました");
     }
 });
-
-function uploadFileToGCS(upFile: Buffer, onFinish: (fileName: string) => void, onError: (err: Error) => void) {
-    const fileName = uuidv4() + '.mp3';
-    const storage = new Storage();
-    const stream = storage.bucket(EnvironmentVariable.bucketName).file(fileName).createWriteStream({
-        metadata: {
-            contentType: 'audio/mp3',
-        },
-        resumable: false
-    });
-    stream.on('error', (err) => {
-        onError(err);
-    });
-    stream.on('finish', () => {
-        console.log('<GCS>upload file');
-        onFinish(fileName);
-    });
-    stream.end(upFile);
-}
 
 async function sendMail(transcript: string | null, toAddress: string, mailText: string) {
     if (EnvironmentVariable.fromAddress === "") {
